Hand Assemble demo program. Create mem2mif translation script
Quartus has much better support for mif files (allegedly), transistioned to that to see if it fixed issues compiling. Reworked memory read in program.v to support mif file. It fixed the errors I was getting about memory ports that didn't exist, but that was not the whole issue.

Minor bugs in assembly were found and fixed. A few minor optimizations were added.